requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via okdata-data-collectors (setup.py)
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via starlette
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
aws-xray-sdk==2.12.1
    # via okdata-data-collectors (setup.py)
boto3==1.34.41
    # via okdata-data-collectors (setup.py)
botocore==1.34.41
    # via
    #   aws-xray-sdk
    #   boto3
    #   s3transfer
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
deprecation==2.1.0
    # via python-keycloak
ecdsa==0.18.0
    # via python-jose
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.6
    # via
    #   anyio
    #   requests
    #   yarl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.21.1
    # via okdata-sdk
jsonschema-specifications==2023.12.1
    # via jsonschema
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
okdata-aws==2.0.0
    # via okdata-data-collectors (setup.py)
okdata-sdk==3.1.0
    # via okdata-aws
packaging==23.2
    # via deprecation
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pydantic==2.6.1
    # via okdata-aws
pydantic-core==2.16.2
    # via pydantic
pyjwt==2.8.0
    # via okdata-sdk
python-dateutil==2.8.2
    # via botocore
python-jose==3.3.0
    # via
    #   okdata-sdk
    #   python-keycloak
python-keycloak==3.8.3
    # via okdata-sdk
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   okdata-aws
    #   okdata-sdk
    #   python-keycloak
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via python-keycloak
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via python-jose
s3transfer==0.10.0
    # via boto3
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via anyio
starlette==0.37.1
    # via okdata-aws
structlog==24.1.0
    # via okdata-aws
typing-extensions==4.9.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==1.26.18
    # via
    #   botocore
    #   okdata-sdk
    #   requests
wrapt==1.16.0
    # via aws-xray-sdk
yarl==1.9.4
    # via aiohttp
